+++ src/index.tsx
@@ -1,21 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import * as firebase from 'firebase/app'
-
+import initializeFirebase from 'settings/firebase'
 import App from './App'
 import './index.css'
 
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_CONFIG_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_CONFIG_AUTH_DOMAIN,
-  databaseURL: process.env.REACT_APP_FIREBASE_CONFIG_DATABASE_URL,
-  projectId: process.env.REACT_APP_FIREBASE_CONFIG_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_CONFIG_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_CONFIG_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_CONFIG_APP_ID,
-}
-
-firebase.initializeApp(firebaseConfig)
+initializeFirebase()
 
 ReactDOM.render(
   <React.StrictMode>
