+++ src/settings/firebase.ts
@@ -1,0 +1,17 @@
+import * as firebase from 'firebase/app'
+
+const firebaseConfig = {
+  apiKey: process.env.REACT_APP_FIREBASE_CONFIG_API_KEY,
+  authDomain: process.env.REACT_APP_FIREBASE_CONFIG_AUTH_DOMAIN,
+  databaseURL: process.env.REACT_APP_FIREBASE_CONFIG_DATABASE_URL,
+  projectId: process.env.REACT_APP_FIREBASE_CONFIG_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_FIREBASE_CONFIG_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_FIREBASE_CONFIG_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_FIREBASE_CONFIG_APP_ID,
+}
+
+const initializeFirebase = (): void => {
+  firebase.initializeApp(firebaseConfig)
+}
+
+export default initializeFirebase
